fix: make the history page work with lots of events
Adds a pagination of 100 elements and a load more button to display the next batch

(cherry picked from commit 766d28e997457d7bd083b89c263b564d8a438240)

+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/rest/api/service/impl/ApiServiceImpl.java
@@ -1514,8 +1514,8 @@ public class ApiServiceImpl extends AbstractService implements ApiService {
                 executionContext,
                 Arrays.asList(PUBLISH_API, EventType.UNPUBLISH_API),
                 properties,
-                0,
-                0,
+                (long) 0,
+                (long) 0,
                 0,
                 1,
                 singletonList(executionContext.getEnvironmentId())
